+++ projects/packages/jetpack-mu-wpcom/src/class-jetpack-mu-wpcom.php
@@ -14,7 +14,7 @@ namespace Automattic\Jetpack;
  */
 class Jetpack_Mu_Wpcom {
 
-	const PACKAGE_VERSION = '0.2.3-alpha';
+	const PACKAGE_VERSION = '1.1.2';
 	const PKG_DIR         = __DIR__ . '/../';
 
 	/**
@@ -32,6 +32,7 @@ class Jetpack_Mu_Wpcom {
 
 		// Coming Soon feature.
 		add_action( 'plugins_loaded', array( __CLASS__, 'load_coming_soon' ) );
+		add_action( 'plugins_loaded', array( __CLASS__, 'load_wpcom_rest_api_endpoints' ) );
 
 		/**
 		 * Runs right after the Jetpack_Mu_Wpcom package is initialized.
@@ -49,6 +50,9 @@ class Jetpack_Mu_Wpcom {
 		 * On WoA sites, users may be using non-symlinked older versions of the FSE plugin.
 		 * If they are, check the active version to avoid redeclaration errors.
 		 */
+		if ( ! function_exists( 'is_plugin_active' ) ) {
+			require_once ABSPATH . 'wp-admin/includes/plugin.php';
+		}
 		$invalid_fse_version_active = is_plugin_active( 'full-site-editing/full-site-editing-plugin.php' ) && version_compare( get_plugin_data( WP_PLUGIN_DIR . '/full-site-editing/full-site-editing-plugin.php' )['Version'], '3.56084', '<' );
 		if ( $invalid_fse_version_active ) {
 			return;
@@ -59,6 +63,26 @@ class Jetpack_Mu_Wpcom {
 			apply_filters( 'a8c_enable_public_coming_soon', false )
 		) {
 			require_once __DIR__ . '/features/coming-soon/coming-soon.php';
+		}
+	}
+
+	/**
+	 * Load WP REST API plugins for wpcom
+	 */
+	public static function load_wpcom_rest_api_endpoints() {
+		if ( ! function_exists( 'wpcom_rest_api_v2_load_plugin' ) ) {
+			return;
+		}
+
+		// We don't use `wpcom_rest_api_v2_load_plugin_files` because it operates inconsisently.
+		$plugins = glob( __DIR__ . '/features/wpcom-endpoints/*.php' );
+
+		if ( ! is_array( $plugins ) ) {
+			return;
+		}
+
+		foreach ( array_filter( $plugins, 'is_file' ) as $plugin ) {
+			require_once $plugin;
 		}
 	}
 }