+++ projects/packages/videopress/src/client/admin/components/video-details/stories/index.tsx
@@ -16,6 +16,8 @@ const VideoDetailsTemplate: ComponentStory< typeof VideoDetails > = VideoDetails
 
 export const Default = VideoDetailsTemplate.bind( {} );
 Default.args = {
+	guid: 'ezoR6kzb',
 	filename: 'video-thumbnail.png',
 	src: 'https://videos.files.wordpress.com/fx123456B/video-thumbnail.mov',
+	isPrivate: false,
 };