+++ projects/packages/videopress/src/client/components/timestamp-control/types.ts
@@ -1,0 +1,11 @@
+export type TimestampInputProps = {
+	value: number;
+	max?: number;
+	fineAdjustment?: number;
+	onChange?: ( ms: number ) => void;
+};
+
+export type TimestampControlProps = TimestampInputProps & {
+	wait?: number;
+	onDebounceChange?: ( ms: number ) => void;
+};