+++ projects/plugins/jetpack/_inc/blogging-prompts.php
@@ -48,6 +48,52 @@ function jetpack_setup_blogging_prompt_response( $post_id ) {
 add_action( 'wp_insert_post', 'jetpack_setup_blogging_prompt_response' );
 
 /**
+ * When a published posts answers a blogging prompt, store the prompt id in the post meta.
+ *
+ * @param int          $post_id     Post ID.
+ * @param WP_Post      $post        Post object.
+ * @param bool         $update      Whether this is an existing post being updated.
+ * @param null|WP_Post $post_before Null for new posts, the WP_Post object prior
+ *                                  to the update for updated posts.
+ */
+function jetpack_mark_if_post_answers_blogging_prompt( $post_id, $post, $update, $post_before ) {
+	if ( ! $post instanceof WP_Post ) {
+		return;
+	}
+
+	$post_type    = isset( $post->post_type ) ? $post->post_type : null;
+	$post_content = isset( $post->post_content ) ? $post->post_content : null;
+
+	if ( 'post' !== $post_type || ! $post_content ) {
+		return;
+	}
+
+	$new_status = isset( $post->post_status ) ? $post->post_status : null;
+	$old_status = $post_before && isset( $post_before->post_status ) ? $post_before->post_status : null;
+
+	// Make sure we are publishing a post, and it's not already published.
+	if ( 'publish' !== $new_status || 'publish' === $old_status ) {
+		return;
+	}
+
+	$blocks = parse_blocks( $post->post_content );
+	foreach ( $blocks as $block ) {
+		if ( 'jetpack/blogging-prompt' === $block['blockName'] ) {
+			$prompt_id      = isset( $block['attrs']['promptId'] ) ? absint( $block['attrs']['promptId'] ) : null;
+			$has_prompt_tag = has_tag( 'dailyprompt', $post ) || ( $prompt_id && has_tag( "dailyprompt-{$prompt_id}", $post ) );
+
+			if ( $prompt_id && $has_prompt_tag && count( $blocks ) > 1 ) {
+				update_post_meta( $post->ID, '_jetpack_blogging_prompt_key', $prompt_id );
+			}
+
+			break;
+		}
+	}
+}
+
+add_action( 'wp_after_insert_post', 'jetpack_mark_if_post_answers_blogging_prompt', 10, 4 );
+
+/**
  * Utility functions.
  */